pythonServer/requirements.txt
fastapi
uvicorn
python-dotenv
certifi
discord.py
beautifulsoup4
sqlalchemy
psycopg2-binary
openai
